update compelted TD functionality
